# src/main/java/com/nmamou/contacts/web/ContactController.java
package com.nmamou.contacts.web;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

import com.nmamou.contacts.pojo.Contact;
import com.nmamou.contacts.service.ContactService;

// SpringDoc OpenAPI UI mvn repository dependency
// <version>1.6.9</version>
// OpenApi json file generation
// http://localhost:8000/v3/api-docs

// http://localhost:8000/swagger-ui/index.html
// Then View The Api Ui documentation using swagger

@RestController
public class ContactController {
  @Autowired
  private ContactService contactService;

  @GetMapping("/contact/{id}")
  public ResponseEntity<Contact> getContact(@PathVariable String id){
    Contact contact = contactService.getContactById(id);
    return new ResponseEntity<>(contact, HttpStatus.OK);
  }

  @GetMapping("/contacts")
  public ResponseEntity<List<Contact>> getContacts(){
    List<Contact> contacts = contactService.getContacts();
    return new ResponseEntity<>(contacts, HttpStatus.OK);
  }


  @PostMapping("/contact")
  public ResponseEntity<HttpStatus> createContact(@RequestBody Contact contact){
    contactService.saveContact(contact);
   return new ResponseEntity<>(HttpStatus.CREATED);
  }

  @PutMapping("/contact/{id}")
  public ResponseEntity<Contact> updateContact(@PathVariable String id, @RequestBody Contact contact){
    contactService.updateContact(id, contact);
    return new ResponseEntity<Contact>(contact, HttpStatus.CREATED);
  }
  
  @DeleteMapping("/delete/{id}/contact")
  public ResponseEntity<HttpStatus> deleteContact(@PathVariable String id){
    contactService.deleteContact(id);
    return new ResponseEntity<>(HttpStatus.OK);
  }

}